requirements.txt
fastapi
uvicorn
jinja2
python-multipart
psycopg2-binary
bcrypt
itsdangerous
python-dotenv